Modified player ship to stay within screen bounds.

+++ app/src/main/java/com/jacstuff/spacearmada/view/fragments/game/GameView.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface GameView {
 
-    void updateShip(int x, int y);
+    void updateShip(float x, float y);
     void updateEnemyShip(int x, int y);
     void updateStars(List<Point> starCoordinates);
 }
